requirements.txt
Flask==2.3.2
openai==1.30.1
python-dotenv==1.0.1
